third_party/blink/web_tests/external/wpt/xhr/setrequestheader-case-insensitive-expected.txt

This is a testharness.js-based test.
[FAIL] XMLHttpRequest: setRequestHeader() - headers that differ in case 1
  assert_equals: expected ["content-TYPE"] but got ["content-TYPE"]
Harness: the test ran to completion.
